Updated README. Changed panda config to be driven by our custom preset.  Added Claude.md.
