+++ src/recipes/index.ts
@@ -2,18 +2,20 @@ export { buttonRecipe, iconButtonRecipe } from './button';
 export { inputRecipe } from './input';
 export { textareaRecipe } from './textarea';
 export { textRecipe, headingRecipe, linkRecipe, labelRecipe } from './text';
-export { checkBoxRecipe } from './checkbox';
 export { spinnerRecipe } from './spinner';
 export { dividerRecipe } from './divider';
 export { preRecipe, codeRecipe } from './code';
 export { boxRecipe } from './box';
-export { radioRecipe } from './radio';
 export { toggleRecipe } from './toggle';
 export { textinputRecipe } from './textinput';
-export { radioInputRecipe } from './radioinput';
 export { toggleInputRecipe } from './toggleinput';
 export { cardRecipe } from './card';
-export { tooltipRecipe } from './tooltip';
 export { breadcrumbsRecipe } from './breadcrumbs';
 export { tagRecipe } from './tag';
+export { radioInputRecipe } from './radioinput';
+
+// Slot Recipes
+export { checkBoxRecipe } from './checkbox';
+export { radioRecipe } from './radio';
+export { tooltipRecipe } from './tooltip';
 export { menuRecipe } from './menu';
